create dashboard folder, dashboard index.css file, moved dashboard-related css to index.css file in dashboard
