+++ src/components/Dashboard/index.css
@@ -1,0 +1,167 @@
+/* Dashboard Box */
+
+.box {
+    background: #FFFFFF;
+    box-shadow: 0px 2px 16px rgba(128, 72, 91, 0.05), 0px 2px 4px rgba(128, 72, 91, 0.1);
+    border-radius: 8px;
+}
+
+.wrap {
+    padding: 40px 80px;
+}
+
+/* Dashboard */
+
+.line-chart{
+    position: absolute;
+    width: 778px;
+    height: 441px;
+    left: 80px;
+    top: 362px;
+    background: #FFFFFF;
+    box-shadow: 0px 2px 16px rgba(128, 72, 91, 0.05), 0px 2px 4px rgba(128, 72, 91, 0.1);
+    border-radius: 8px;
+    }
+
+.keystats{
+    position: absolute;
+    left: 79.53%;
+    right: 9.77%;
+    top: 28.29%;
+    bottom: 70.5%;
+    font-size: 26px;
+    line-height: 33px;
+    display: flex;
+    align-items: center;
+}
+
+.totalnumberdonors{
+    width: 294px;
+    height: 112px;
+    left: 906px;
+    top: 362px;
+    font-size: 26px;
+    line-height: 17px;
+    color: #252631
+}
+
+.totaldonationamt{
+    font-size: 26px;
+    line-height: 17px;
+    color: #252631
+}
+
+.keystatslabel{
+    font-size: 14px;
+    line-height: 22px;  
+    color: #9FA2B4
+}
+
+.steps{
+    position: absolute;
+    left: 87.29%;
+    right: -64.09%;
+    top: 23.2%;
+    bottom: 14.64%;
+    transform: matrix(-1, 0, 0, 1, 0, 0);
+    height: 278px;
+    width: 225px;
+    }
+
+/* BySource Component */
+.legend {
+   width: 12px;
+   height: 12px;
+   border-radius: 50%; 
+}
+
+.legend-container{
+   margin: 30px 48px auto 0;
+}
+
+.brightorange{
+   background: #FFA001;
+}
+
+.pink{
+   background: #FE5366;
+}
+
+.lightpink{
+   background:  #FF9A85;
+}
+
+.darkpurple{
+   background: #80485B;
+}
+
+.lilac{
+   background: #CB87B0;
+}
+
+.flex{
+   display: flex;
+   align-items: baseline;
+}
+
+.bysource-crownvalue{
+   font-size: 24px;
+   margin-left: 12.96px;
+}
+
+.bysource-value{
+   font-size: 24px; 
+   margin-left: 29.99px
+}
+.bysource-name{
+   margin-left: 16px;
+   line-height: 18px;
+   font-size: 14px;
+   color: #9FA2B4;
+   letter-spacing: 0.3px;
+}
+
+.bysource-row1{
+   width: 382px;
+   height: 110px;
+   padding-top: 30px;   
+}
+
+.bysource-row{
+   width: 191px;
+   height: 110px;
+   padding-top: 25px;   
+}
+
+.piechart-holder{
+     position: relative; 
+     display: flex;
+     justify-content: center;
+     align-items: center;
+    
+}
+
+.piechart-img{
+    position: absolute; 
+    top: 50% ;
+    left: 50%;
+    border-radius: 100%;
+    width : 128px;
+    height: 128px; 
+    box-shadow: 0px 0px 24px rgba(128, 72, 91, 0.08);
+    transform: translate(-50%, -50%);
+    display: flex;
+    justify-content: center;
+    
+}
+
+
+.chart{
+    width: 778px;
+    height: 398px;
+    border-radius: 8px;
+    display: flex;
+    margin: 2rem 2rem;
+    background: #FFFFFF;
+    box-shadow: 0px 2px 16px rgba(128, 72, 91, 0.05), 0px 2px 4px rgba(128, 72, 91, 0.1);
+}
